src/frontend/pyproject.toml
[project]
name = "portfolio-tuner-frontend"
version = "1.5.0"
description = "Streamlit frontend dashboard for Portfolio Tuner API"
requires-python = ">=3.13"
dependencies = [
    "humanize>=4.14.0",
    "plotly>=6.5.0",
    "requests>=2.32.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.1",
    "streamlit>=1.53.0",
    "streamlit-autorefresh>=1.0.1",
    "tzlocal>=5.3.1",
]
